prepared framework around sdf solver. It works like this:
1. You have a source PGraphics, that you set as a uniform to sdfSolver
(this works since we're not rendering to the source PGraphics).
2. use #define PROCESSING_TEXTURE_SHADER in sdfSolver.glsl. That gives
us access to:
"uniform sampler2D texture;" (this is the old ppixels).
"varying vec4 vertColor;"
"varying vec4 vertTexCoord;"

and allows us to invoke the shader by calling "sdf.filter(sdfSolver)",
instead of the previous hackier sdf.rect().

This also lets us get rid of having to pass a resolution uniform, since
we only need it to compute uv, which is in vertTexCoord.st.

So, in total, to get this running, you need:

1. use #define PROCESSING_TEXTURE_SHADER
2. to grab old frame, use texture uniform
3. pass new data as a uniform
4. invoke compute with sdf.filter()
5. get uv with vertTexCoord.st